+++ aoc_drugsale/fxmanifest.lua
@@ -1,3 +1,8 @@
+shared_script "@ReaperV4/imports/bypass.lua"
+shared_script "@ReaperV4/imports/bypass_s.lua"
+shared_script "@ReaperV4/imports/bypass_c.lua"
+lua54 "yes" -- needed for Reaper
+
 fx_version 'cerulean'
 game 'gta5'
 
@@ -13,7 +18,11 @@ shared_scripts {
 }
 
 client_script 'client.lua'
-server_script 'server.lua'
+server_script {
+      '@oxmysql/lib/MySQL.lua',
+      'server.lua'
+}
+
 
 files {
     'data/reputation.json',